beat buffer now uses doubles, data jockey now uses libsyck and libyamlcpp, beat buffer reads yaml data.
git-svn-id: svn+ssh://yew/home/alex/code-repo/projects/datajockey/code/qtgui@703 be07c712-5235-0410-95d4-c99efef06bb2

+++ include/buffer.hpp
@@ -36,7 +36,7 @@ namespace DataJockey {
 
 	class BeatBuffer : public Object {
 		private:
-			std::vector<float> mBeatBuffer;
+			std::vector<double> mBeatBuffer;
 			unsigned int mId;
 			static unsigned int mIdCnt;
 			void load(std::string beatDataLocation) throw(std::runtime_error);
